+++ songname-server/src/main/java/com/konfyrm/songname/players/dto/GetPlayersResponse.java
@@ -11,9 +11,9 @@ public class GetPlayersResponse {
     public class PlayerDto {
         private UUID uuid;
         private String name;
-        private int score;
+        private float score;
 
-        public PlayerDto(UUID uuid, String name, int score) {
+        public PlayerDto(UUID uuid, String name, float score) {
             this.uuid = uuid;
             this.name = name;
             this.score = score;
@@ -27,7 +27,7 @@ public class GetPlayersResponse {
             return name;
         }
 
-        public int getScore() {
+        public float getScore() {
             return score;
         }
     }
